add-message-after-id attribute added
With the add-message-after-id attribute its possible to specify the element where the error messages will be added after. This may be useful in situations when adding the message after the normal element breaks the design of the form. E.g. when using bootstrap's input-group-addon class on a div element directly after the input element, the layout breaks.