move all dependencies to the devDependencies
This directive does not need any dependencies for doing its job in an Angular app.
Now when we install `angular-validator` via NPM all packages under actual "dependencies" comes with this package - but we don't need them. In my computer the installation even fails because incompatibility of required "dependencies" with my NodeJS. Please move all "dependencies" to the "devDependencies".